added manage listings functionality and UI

+++ app/Http/Controllers/ListingController.php
@@ -52,6 +52,10 @@ class ListingController extends Controller
   }
 
   public function update(Request $request, Listing $listing) {
+    if($listing->user_id != auth()->id()) {
+      abort(403, 'Unauthorized action');
+    }
+
     $formFields = $request->validate([
       'title' => 'required',
       'company' => 'required',
@@ -71,7 +75,15 @@ class ListingController extends Controller
   }
 
   public function destroy(Listing $listing) {
+    if($listing->user_id != auth()->id()) {
+      abort(403, 'Unauthorized action');
+    }
+
     $listing->delete();
     return redirect('/')->with('message', 'Listing deleted successfully!');
+  }
+
+  public function manage() {
+    return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
   }
 }
